MP_SIL: back port HILSIM plugin fixes, add SIL simulation of turning RC Radio On/Off
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_SIL@2105 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ Tools/HILSIM_XPlane/SerialIO.h
@@ -3,7 +3,7 @@
 void OpenComms(void);
 void CloseComms(void);
 
-void StartServer(long PortNum);
+void StartServer(uint16_t PortNum);
 void StopServer(void);
 
 void SendToComPort(unsigned long ResponseLength, unsigned char *Buffer);
