+++ Tools/HILSIM_XPlane/UDBSocket.h
@@ -32,7 +32,7 @@ typedef struct UDBSocket_t *UDBSocket;
 // UDBSocketUDPClient:		specify type, UDP_port, UDP_host
 // UDBSocketUDPServer:		specify type, UDP_port
 // UDBSocketSerial:			specify type, serial_port, serial_baud
-UDBSocket UDBSocket_init(UDBSocketType type, long UDP_port, char *UDP_host, char *serial_port, long serial_baud);
+UDBSocket UDBSocket_init(UDBSocketType type, uint16_t UDP_port, char *UDP_host, char *serial_port, long serial_baud);
 
 void UDBSocket_close(UDBSocket socket);
